refactor prompts directory
