+++ src/text/prompts/content-structure-summaries/bullet-points.ts
@@ -1,0 +1,8 @@
+export const bulletPoints = {
+  instruction: `- Write a bullet point list summarizing the transcript.\n`,
+  example: `## Bullet Point Summary
+
+    - A concise summary of a chapter's content in bullet point list form.
+    - It begins by introducing the main topic or theme of the chapter, providing context for the reader.
+    - The summary then outlines key points or arguments presented in the chapter\n`,
+}
