+++ src/text/prompts/content-structure-summaries/chapter-titles-and-quotes.ts
@@ -1,0 +1,11 @@
+export const chapterTitlesAndQuotes = {
+  instruction: `- Create chapter titles based on the topics discussed throughout the transcript.
+  - Include timestamps for when these chapters begin.
+  - Chapters should be roughly 3-6 minutes long.
+  - Under each chapter title, include the most representative quote from that chapter.\n`,
+  example: `## Chapters with Quotes
+    
+    ### 00:00:00 - Introduction to Web Development
+    
+    > "Web development is constantly evolving, driven by new technologies and user expectations."\n`,
+}
