+++ src/text/prompts/content-structure-summaries/chapter-titles.ts
@@ -1,0 +1,12 @@
+export const chapterTitles = {
+  instruction: `- Create chapter titles based on the topics discussed throughout the transcript.
+  - Include only starting timestamps in exact HH:MM:SS format, always using two digits each for hours, minutes, and seconds.
+  - Chapters should be roughly 3-6 minutes long.`,
+  example: `## Chapters
+
+    ### 00:00:00 - Introduction and Overview
+    ### 00:03:12 - The History of Web Development
+    ### 00:07:45 - The Rise of JavaScript Frameworks
+    ### 00:12:30 - Server-Side Technologies
+    ### 00:18:00 - The Future of Web Development`,
+}
