+++ src/text/prompts/content-structure-summaries/faq.ts
@@ -1,0 +1,8 @@
+export const faq = {
+  instruction: `- Include a list of 5-10 frequently asked questions and answers based on the transcript.
+  - Ensure questions and answers cover all major sections of the content.\n`,
+  example: `## FAQ
+
+    Q: What are the three main components of the Jamstack?
+    A: JavaScript, APIs, and markup.\n`,
+}
