+++ src/text/prompts/content-structure-summaries/long-summary.ts
@@ -1,0 +1,7 @@
+export const longSummary = {
+  instruction: `- Write a one paragraph summary of the transcript.
+  - The one paragraph summary should be approximately 600-1200 characters (roughly 100-200 words).\n`,
+  example: `## Episode Summary
+
+    A concise summary of a chapter's content, typically ranging from 600 to 1200 characters or approximately 100 to 200 words. It begins by introducing the main topic or theme of the chapter, providing context for the reader. The summary then outlines key points or arguments presented in the chapter, touching on major concepts, theories, or findings discussed. It may briefly mention methodologies used or data analyzed, if applicable. The paragraph also highlights any significant conclusions or implications drawn from the chapter's content. Throughout, it maintains a balance between providing enough detail to give readers a clear understanding of the chapter's scope and keeping the information general enough to apply to various subjects. This summary style efficiently conveys the essence of the chapter's content, allowing readers to quickly grasp its main ideas and decide if they want to dig deeper into the full text.\n`,
+}
