+++ src/text/prompts/content-structure-summaries/medium-chapters.ts
@@ -1,0 +1,13 @@
+export const mediumChapters = {
+  instruction: `- Create chapter titles and one-paragraph descriptions based on the topics discussed.
+  - Include only starting timestamps in exact HH:MM:SS format, always using two digits each for hours, minutes, and seconds.
+  - Chapters should be roughly 3-6 minutes long.
+  - Write descriptions of about 50 words each.
+  - Ensure chapters cover the entire content, clearly noting the last timestamp (HH:MM:SS), indicating total duration.
+  - Descriptions should flow naturally from the content.`,
+  example: `## Chapters
+
+    ### 00:00:00 - Introduction and Overview
+
+    Introduces the key themes and concepts explored in the episode, briefly outlining the significant points and their relevance to broader discussions. This foundation helps listeners grasp the scope and importance of the subsequent content and prepares them for deeper exploration of each topic.`,
+}
