+++ src/text/prompts/content-structure-summaries/questions.ts
@@ -1,0 +1,24 @@
+export const questions = {
+  instruction: `- Include a list of 10 questions to check the listeners' comprehension of the material.
+  - Ensure questions cover all major sections of the content.
+  - Ensure the questions are correct, emphasize the right things, and aren't redundant.
+  - Do not say things like "the instructor describes" or "according to the lesson," assume that all the questions relate to the lesson told by the instructor.
+  - The first five questions should be beginner level questions and the last five should be expert level questions.\n`,
+  example: `## Questions to Check Comprehension
+
+    ### Beginner Questions
+
+    1. What are the three main components of the modern web development stack?
+    2. How has the role of JavaScript evolved in web development over the past decade?
+    3. What are the key differences between React and Vue.js?
+    4. Why is server-side rendering beneficial for web applications?
+    5. What is the purpose of a RESTful API in full-stack development?
+
+    ### Expert Questions
+
+    6. How does Node.js differ from traditional server-side languages like PHP or Python?
+    7. What are the main considerations when choosing a database for a web application?
+    8. How do containerization technologies like Docker impact web development and deployment?
+    9. What role does responsive design play in modern web development?
+    10. How can developers ensure the security of user data in web applications?\n`,
+}
