+++ src/text/prompts/content-structure-summaries/quotes.ts
@@ -1,0 +1,10 @@
+export const quotes = {
+  instruction: `- Select the five most important and impactful quotes from the transcript.\n`,
+  example: `## Important Quotes
+    
+    1. "First important quote from the episode."
+    2. "Second important quote from the episode."
+    3. "Third important quote from the episode."
+    4. "Fourth important quote from the episode."
+    5. "Fifth important quote from the episode."\n`,
+}
