+++ src/text/prompts/content-structure-summaries/short-chapters.ts
@@ -1,0 +1,12 @@
+export const shortChapters = {
+  instruction: `- Create chapter titles and one-sentence descriptions based on the topics discussed.
+  - Include only starting timestamps in exact HH:MM:SS format, always using two digits each for hours, minutes, and seconds.
+  - Chapters should be roughly 3-6 minutes long.
+  - Ensure chapters cover the entire content, clearly noting the last timestamp (HH:MM:SS), indicating total duration.
+  - Descriptions should flow naturally from the content.`,
+  example: `## Chapters
+
+    ### 00:00:00 - Introduction and Episode Overview
+
+    Briefly introduces the episode's main themes, setting the stage for detailed discussions ahead.`,
+}
