+++ src/text/prompts/content-structure-summaries/short-summary.ts
@@ -1,0 +1,7 @@
+export const shortSummary = {
+  instruction: `- Write a one sentence description of the transcript.
+  - The one sentence description shouldn't exceed 180 characters (roughly 30 words).\n`,
+  example: `## Episode Description
+
+    One sentence description of the transcript that encapsulates the content contained in the file but does not exceed roughly 180 characters (or approximately 30 words).\n`,
+}
