+++ src/text/prompts/content-structure-summaries/summary.ts
@@ -1,0 +1,12 @@
+export const summary = {
+  instruction: `- Write a one-sentence description of the transcript and a one-paragraph summary.
+  - The one-sentence description shouldn't exceed 180 characters (roughly 30 words).
+  - The one-paragraph summary should be approximately 600-1200 characters (roughly 100-200 words).`,
+  example: `## Episode Description
+
+    One sentence description encapsulating the content within roughly 180 characters.
+
+    ## Episode Summary
+
+    A concise summary of the transcript, typically 600-1200 characters or about 100-200 words, highlighting main topics, significant points, methods, conclusions, and implications.`,
+}
