+++ src/text/prompts/content-structure-summaries/takeaways.ts
@@ -1,0 +1,8 @@
+export const takeaways = {
+  instruction: `- Include three key takeaways the listener should get from the episode.\n`,
+  example: `## Key Takeaways
+
+    1. Full-stack development requires a broad understanding of both client-side and server-side technologies, enabling developers to create more cohesive and efficient web applications.
+    2. Modern front-end frameworks like React and Vue.js have revolutionized UI development, emphasizing component-based architecture and reactive programming paradigms.
+    3. Backend technologies like Node.js and cloud services have made it easier to build scalable, high-performance web applications, but require careful consideration of security and data management practices.\n`,
+}
