+++ src/text/prompts/content-structure-summaries/titles.ts
@@ -1,0 +1,12 @@
+export const titles = {
+  instruction: `- Write 5 potential titles for the video.
+  - The first two titles should be very short and have no subtitle.
+  - The last three titles can be longer and include subtitles.`,
+  example: `## Potential Titles
+
+    1. Title Hard
+    2. Title Harder
+    3. Title Hard with a Vengeance
+    4. Title Hard IV: Live Free or Title Hard
+    5. Title Hard V: A Good Day to Die Hard`,
+}
