+++ src/text/prompts/creative-entertainment/country-song.ts
@@ -1,0 +1,9 @@
+export const countrySong = {
+  instruction: `- Write a heartfelt, storytelling country song with simple yet emotionally charged lyrics.
+    - Include themes of life, love, and the struggles of everyday people.
+    - Structure the song with verses, choruses, and a bridge.
+    - Offer the song three potential titles.\n`,
+  example: `## Song
+    
+    Lyrics to the song.\n`
+}
