+++ src/text/prompts/creative-entertainment/rap-song.ts
@@ -1,0 +1,9 @@
+export const rapSong = {
+  instruction: `- Write a highly complex, multi-syllabic rhyming, Eminem inspired rap based on this transcript.
+    - Do not rhyme any words with themselves.
+    - Give it a basic song structure with verses, choruses, and a bridge.
+    - Give the song three potential titles.\n`,
+  example: `## Song
+    
+    Lyrics to the song.\n`
+}
