+++ src/text/prompts/creative-entertainment/rock-song.ts
@@ -1,0 +1,9 @@
+export const rockSong = {
+  instruction: `- Write a high-energy, anthemic rock song with powerful imagery and impactful, multi-layered lyrics.
+    - Use metaphors and vivid language to convey a sense of rebellion or freedom.
+    - Structure the song with verses, choruses, and a bridge.
+    - Provide the song with three potential titles.\n`,
+  example: `## Song
+    
+    Lyrics to the song.\n`
+}
