+++ src/text/prompts/index.ts
@@ -1,0 +1,47 @@
+import { titles } from './content-structure-summaries/titles'
+import { summary } from './content-structure-summaries/summary'
+import { shortSummary } from './content-structure-summaries/short-summary'
+import { longSummary } from './content-structure-summaries/long-summary'
+import { bulletPoints } from './content-structure-summaries/bullet-points'
+import { quotes } from './content-structure-summaries/quotes'
+import { chapterTitlesAndQuotes } from './content-structure-summaries/chapter-titles-and-quotes'
+import { chapterTitles } from './content-structure-summaries/chapter-titles'
+import { faq } from './content-structure-summaries/faq'
+import { shortChapters } from './content-structure-summaries/short-chapters'
+import { mediumChapters } from './content-structure-summaries/medium-chapters'
+import { longChapters } from './content-structure-summaries/long-chapters'
+import { takeaways } from './content-structure-summaries/takeaways'
+import { keyMoments } from './content-structure-summaries/key-moments'
+import { questions } from './content-structure-summaries/questions'
+import { x } from './marketing-social-media/x'
+import { facebook } from './marketing-social-media/facebook'
+import { linkedin } from './marketing-social-media/linkedin'
+import { blog } from './marketing-social-media/blog'
+import { rapSong } from './creative-entertainment/rap-song'
+import { rockSong } from './creative-entertainment/rock-song'
+import { countrySong } from './creative-entertainment/country-song'
+
+export const sections = {
+  titles,
+  summary,
+  shortSummary,
+  longSummary,
+  bulletPoints,
+  quotes,
+  chapterTitlesAndQuotes,
+  x,
+  facebook,
+  linkedin,
+  chapterTitles,
+  shortChapters,
+  mediumChapters,
+  longChapters,
+  takeaways,
+  questions,
+  faq,
+  blog,
+  rapSong,
+  rockSong,
+  countrySong,
+  keyMoments,
+}
