+++ src/text/prompts/marketing-social-media/blog.ts
@@ -1,0 +1,14 @@
+export const blog = {
+  instruction: `- Generate a blog outline and first draft for a blog post based on this piece of content.
+    
+    - Make sure the blog post is at least 750 words.\n`,
+  example: `## Blog Outline
+
+    1. Part 1
+    2. Part 2
+    3. Part 3
+    
+    ### Blog First Draft
+    
+    First draft of a blog.\n`,
+}
