+++ src/text/prompts/marketing-social-media/facebook.ts
@@ -1,0 +1,8 @@
+export const facebook = {
+  instruction: `- Write an engaging Facebook post summarizing the key themes of the transcript.
+  - Keep it conversational and suitable for broad audiences.
+  - Posts should ideally be around 100-200 words.\n`,
+  example: `## Facebook Social Post
+    
+    Have you ever wondered how web applications have transformed over the last decade? In our latest episode, we explore this evolution, highlighting the rise of JavaScript frameworks, backend innovations, and future trends that every developer should watch. Tune in to learn more and stay updated!\n`,
+}
