+++ src/text/prompts/marketing-social-media/linkedin.ts
@@ -1,0 +1,8 @@
+export const linkedin = {
+  instruction: `- Write a professional and insightful LinkedIn post based on the transcript.
+  - Highlight key takeaways or actionable insights relevant to professionals.
+  - Aim for 150-300 words.\n`,
+  example: `## LinkedIn Social Post
+    
+    Today's web developers need to master a blend of skills, from client-side frameworks to robust backend technologies. Our latest content explores how professionals can stay competitive by embracing new tools, adapting to market trends, and continuously learning. What steps are you taking to future-proof your career? Share your thoughts!\n`,
+}
