+++ src/text/prompts/marketing-social-media/x.ts
@@ -1,0 +1,8 @@
+export const x = {
+  instruction: `- Write a concise, engaging social media post optimized for the platform X (formerly Twitter).
+  - Keep the post under 280 characters.
+  - Include hashtags if appropriate.\n`,
+  example: `## X Social Post
+    
+    Web development isn't just coding; it's shaping the future. Dive into the latest trends and stay ahead of the curve. #WebDev #JavaScript\n`,
+}
